+++ src/api/store/useGetStore.ts
@@ -1,40 +1,22 @@
 import { QueryKey, useSuspenseQuery } from '@tanstack/react-query';
 import { useSearchParams } from 'next/navigation';
+import { Store, CostDetail } from '@/api/store';
 
 interface GetStoreRequest {
   page: string;
   limit: string;
 }
 
-export interface Store {
-  store_id: string;
-  store_name: string;
-  store_location: string;
-  store_size: string;
-  store_category: string;
-  store_cost: number;
-  monthly_sales: number;
-  monthly_revenue: number;
-  monthly_cost: number;
-  manager: string;
-  manager_contact: string;
-  created_at: string;
-  updated_at: string;
-  deleted: string;
-  sales_reason: string;
-  store_size_m2?: string;
-  store_img_src_arr?: string[];
-}
-
 export interface GetStoreResponse {
   data: Store[];
+  cost_details: CostDetail[];
 }
 
 const getStore = async ({
   page,
   limit,
 }: GetStoreRequest): Promise<GetStoreResponse> => {
-  const reqUrl = new URL('/api/store', 'http://localhost:3000');
+  const reqUrl = new URL('/api/store', process.env.NEXT_PUBLIC_BASE_URL);
   const parsedPage = String(parseInt(page) - 1);
   reqUrl.searchParams.set('page', parsedPage);
   reqUrl.searchParams.set('limit', limit);
